add vuex module for revenue section
